requirements.txt
selenium==4.16.0
python-telegram-bot
python-dotenv
beautifulsoup4
chromedriver-autoinstaller
